### 2025-12-05 #9546 @editorialbot
[PRE REVIEW]: ESMBenchmarkViz: A Python Toolkit for Interactive Visualization of Earth System Model Evaluation and Benchmarking
**Submitting author:** @lee1043 (<a href="http://orcid.org/0000-0002-0016-7199">Jiwoo Lee</a>)
**Repository:** https://github.com/PCMDI/ESMBenchmarkViz
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.2.2
