+++ docs/paper/paper.md
@@ -7,6 +7,8 @@ tags:
     - climate research
     - climate model
     - climate model evaluation
+    - Earth system model
+    - Earth system model evaluation
     - interactive visualization
     - Taylor Diagram
     - Portrait Plot
@@ -22,17 +24,18 @@ authors:
       affiliation: 1
     - name: Paul Ullrich
       orcid: 0000-0003-4118-4590
-      affiliation: 1
+      affiliation: "1, 2"
 affiliations:
-    - name: Lawrence Livermore National Lab, Livermore, USA
-      index: 1
+    - name: Lawrence Livermore National Lab, United States
+      index: 2
+    - name: UC Davis, United States
 date: 1 December 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-ESMBenchmarkViz is a Python toolkit designed to generate interactive graphs for visualizing statistics and metrics from the evaluation of climate and Earth system models. This toolkit enables researchers to perform more straightforward benchmarking and intercomparison of models.
+`ESMBenchmarkViz` is a Python toolkit designed to generate interactive graphs for visualizing statistics and metrics from the evaluation of climate and Earth system models. This toolkit enables researchers to perform more straightforward benchmarking and intercomparison of models.
 
 # Statement of Need
 
@@ -40,15 +43,23 @@ Earth System Models (ESMs) are essential for understanding and predicting the co
 
 Evaluating ESMs involves comparing model outputs against observational data and other models to assess performance. This process is vital for identifying model strengths and weaknesses, guiding improvements, and deepening our understanding of climate processes. However, ESM evaluations (particularly comprehensive evaluations) often generate large volumes of data (e.g., [@Lee:2019], [@Lee:2021], [@Lee:2024]; [@Ahn:2022]; [@Planton:2021]), which can be challenging to interpret and communicate effectively. Wrangling this data to produce an effective, polished visualization is a cumbersome process that often requires constructing data and label lists, calculating plot coordinates, and making manual adjustments for final elements like color bar position or label orientation.
 
-To address these challenges, we have developed ESMBenchmarkViz, a modern Python library specifically designed for efficient interactive visualization of statistical performance metrics from ESM evaluation and intercomparison. This library leverages the power of Bokeh [@Bokeh:2018] to provide researchers and practitioners with user-friendly interactive tools for dynamic exploration of complex datasets. By enabling real-time manipulation of visualizations, ESMBenchmarkViz facilitates deeper insights into model performance and inter-model comparisons, making it easier to identify outliers and persistent biases.
+To address these challenges, we have developed `ESMBenchmarkViz`, a modern Python library specifically designed for efficient interactive visualization of statistical performance metrics from ESM evaluation and intercomparison. This library leverages the power of Bokeh [@Bokeh:2018] to provide researchers and practitioners with user-friendly interactive tools for dynamic exploration of complex datasets. By enabling real-time manipulation of visualizations, `ESMBenchmarkViz` facilitates deeper insights into model performance and inter-model comparisons, making it easier to identify outliers and persistent biases.
 
-The development of ESMBenchmarkViz originated from the interactive visualization dashboard of the [Program for Climate Model Diagnosis and Intercomparison (PCMDI) Metrics Package](https://pcmdi.github.io/pcmdi_metrics/) [@pcmdi-metrics] [@Lee:2024], showing diverse evaluation metrics for ESMs along with diagnostic information (https://pcmdi.llnl.gov/research/metrics/). We refer to these diagnostics as “dive-down information,” as they enable users to investigate metrics in greater detail.
[…]
