### 2026-01-15 PCMDI/ESMBenchmarkViz#40 @lee1043
JOSS paper review / conda-forge set up
Follow up: https://github.com/openjournals/joss-reviews/issues/9546#issuecomment-3746390917
