### 2026-02-03 #9546 @lee1043
@kthyng thank you for clarification. I was confused because one of previous comment (https://github.com/openjournals/joss-reviews/issues/9546#issuecomment-3793393753) didn't include the "Statement of need" section, so I thought it should be renamed to "State of the field". Do you happen to have entire list of required sections, or can I consider https://github.com/openjournals/joss-reviews/issues/9546#issuecomment-3843379421 as the list?
