### 2026-03-17 #9546 @kthyng
@lee1043
* can you get this on PyPI or conda-forge?
* you have some placeholders in your docs: https://pcmdi.github.io/ESMBenchmarkViz/
