### 2026-03-18 3272038
Enhance documentation for ESMBenchmarkViz package
Expanded documentation for the ESMBenchmarkViz package, detailing its purpose, functionality, and integration with the Bokeh library for interactive visualizations.
