### 2026-03-18 e123cad
Refine documentation for ESMBenchmarkViz package
Updated descriptions for clarity and conciseness throughout the documentation.
