### 2026-03-19 f5b0c2d
Update README with package information table
Added a table with package information including downloads, version, and platforms.
