+++ README.md
@@ -12,6 +12,10 @@
 [![Formatted with black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/python/black)
 
 
+| Name | Downloads | Version | Platforms |
+| --- | --- | --- | --- |
+| [![Conda Recipe](https://img.shields.io/badge/recipe-esmbenchmarkviz-green.svg)](https://anaconda.org/conda-forge/esmbenchmarkviz) | [![Conda Downloads](https://img.shields.io/conda/dn/conda-forge/esmbenchmarkviz.svg)](https://anaconda.org/conda-forge/esmbenchmarkviz) | [![Conda Version](https://img.shields.io/conda/vn/conda-forge/esmbenchmarkviz.svg)](https://anaconda.org/conda-forge/esmbenchmarkviz) | [![Conda Platforms](https://img.shields.io/conda/pn/conda-forge/esmbenchmarkviz.svg)](https://anaconda.org/conda-forge/esmbenchmarkviz) |
+
 ESMBenchmarkViz is a Python-based interactive visualization toolkit built with Bokeh, designed to facilitate the collective evaluation, intercomparison, and benchmarking of Earth System Models (ESMs). The tool provides a user-friendly interface for analyzing model performance, enabling researchers to make informed comparisons and insights into ESM behavior.
 
 ## Developer
