### 2026-03-19 2710844
Enhance installation guide for esmbenchmarkviz
Added detailed installation instructions for esmbenchmarkviz using conda and mamba, including version search and dependency queries.
