### 2026-03-19 bb51a6f
Format installation commands as code blocks
Updated installation instructions to use code blocks for commands.
