+++ docs/installation.rst
@@ -5,47 +5,51 @@ Installation using conda (recommended)
 
 Installing `esmbenchmarkviz` from the `conda-forge` channel can be achieved by adding `conda-forge` to your channels with:
 
-```
-conda config --add channels conda-forge
-conda config --set channel_priority strict
-```
+.. code-block:: bash
+
+    conda config --add channels conda-forge
+    conda config --set channel_priority strict
 
 Once the `conda-forge` channel has been enabled, `esmbenchmarkviz` can be installed with `conda`:
 
-```
-conda install esmbenchmarkviz
-```
+.. code-block:: bash
+
+    conda install esmbenchmarkviz
 
 or with `mamba`:
 
-```
-mamba install esmbenchmarkviz
-```
+.. code-block:: bash
+
+    mamba install esmbenchmarkviz
+
 
 It is possible to list all of the versions of `esmbenchmarkviz` available on your platform with `conda`:
 
-```
-conda search esmbenchmarkviz --channel conda-forge
-```
+.. code-block:: bash
+
+    conda search esmbenchmarkviz --channel conda-forge
+
 
 or with `mamba`:
 
-```
-mamba search esmbenchmarkviz --channel conda-forge
-```
+.. code-block:: bash
+
+    mamba search esmbenchmarkviz --channel conda-forge
+
 
 Alternatively, `mamba repoquery` may provide more information:
 
-```
-# Search all versions available on your platform:
-mamba repoquery search esmbenchmarkviz --channel conda-forge
+.. code-block:: bash
 
-# List packages depending on `esmbenchmarkviz`:
-mamba repoquery whoneeds esmbenchmarkviz --channel conda-forge
+    # Search all versions available on your platform:
+    mamba repoquery search esmbenchmarkviz --channel conda-forge
+    
+    # List packages depending on `esmbenchmarkviz`:
+    mamba repoquery whoneeds esmbenchmarkviz --channel conda-forge
+    
+    # List dependencies of `esmbenchmarkviz`:
+    mamba repoquery depends esmbenchmarkviz --channel conda-forge
 
-# List dependencies of `esmbenchmarkviz`:
-mamba repoquery depends esmbenchmarkviz --channel conda-forge
-```
 
 Installation from the source code
 =================================
