### 2026-03-19 902ba53
Update installation instructions for ESMBenchmarkViz
