### 2026-03-19 #9546 @lee1043
@kthyng we have established conda-forge release and updated the document. Could you please help resume the review process? I am wondering if there is any chance to get the peer-review expedited considering the first submission was made in Dec last year but peer-review has not been started yet, which takes longer than I was hoping for. Thank you for your consideration.
