### 2026-03-19 01da903
Fix affiliation index for authors in paper.md
