### 2026-03-19 36fd0ef
Fix author affiliation and index formatting
Updated author affiliations and index values to use quotes.
