### 2026-03-19 #9546 @lee1043
@kthyng thank you for the follow up. I have revised the affiliation index where the PDF generation error log was pointing to. Could you retry?
