### 2026-04-14 #9546 @lee1043
Just for the reviewers' convenience: 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.09546/joss.09546/10.21105.joss.09546.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.09546/joss.09546/10.21105.joss.09546.pdf) 📄 👈 (https://github.com/openjournals/joss-reviews/issues/9546#issuecomment-4099565746)
