### 2026-04-20 #9546 @mengqi-z
👋 @maruti-iitm @DejaBraz - Would any of you be interested and available to review this submission for the Journal of Open Source Software (JOSS)? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
