### 2026-06-03 bc898ef
Add project dependencies
Added dependencies for numpy, bokeh, matplotlib, and pandas.
