### 2026-06-03 7ccd9ad
add more description and reference for taylor diagram
