+++ _sources/installation.rst.txt
@@ -1,0 +1,114 @@
+.. _installation:
+
+Installation using conda (recommended)
+======================================
+
+Installing `esmbenchmarkviz` from the `conda-forge` channel can be achieved by adding `conda-forge` to your channels with:
+
+.. code-block:: bash
+
+    conda config --add channels conda-forge
+    conda config --set channel_priority strict
+
+Once the `conda-forge` channel has been enabled, `esmbenchmarkviz` can be installed with `conda`:
+
+.. code-block:: bash
+
+    conda install esmbenchmarkviz
+
+or with `mamba`:
+
+.. code-block:: bash
+
+    mamba install esmbenchmarkviz
+
+
+It is possible to list all of the versions of `esmbenchmarkviz` available on your platform with `conda`:
+
+.. code-block:: bash
+
+    conda search esmbenchmarkviz --channel conda-forge
+
+
+or with `mamba`:
+
+.. code-block:: bash
+
+    mamba search esmbenchmarkviz --channel conda-forge
+
+
+Alternatively, `mamba repoquery` may provide more information:
+
+.. code-block:: bash
+
+    # Search all versions available on your platform:
+    mamba repoquery search esmbenchmarkviz --channel conda-forge
+    
+    # List packages depending on `esmbenchmarkviz`:
+    mamba repoquery whoneeds esmbenchmarkviz --channel conda-forge
+    
+    # List dependencies of `esmbenchmarkviz`:
+    mamba repoquery depends esmbenchmarkviz --channel conda-forge
+
+
+Installation from the source code
+=================================
+
+To install from the source code, you will need to clone the repository and install it locally using `pip`. The following steps will guide you through the process.
+
+Prerequisites
+-------------
+
+- Python 3.10 or higher
+- `git <https://github.com/git-guides/install-git>`_ and `pip <https://pip.pypa.io/en/stable/installation/>`_ installed on your system
+- `numpy <https://numpy.org/>`_, `bokeh <https://docs.bokeh.org/en/latest/>`_, `matplotlib <https://matplotlib.org/>`_ libraries installed (These will be installed automatically if you follow the conda environment setup in Step 2)
+- (Optional, but recommended) `conda <https://docs.conda.io/en/latest/>`_ for environment management
+
+Step 1: Clone the Repository
+----------------------------
+
+Clone the repository from `GitHub <https://github.com/PCMDI/ESMBenchmarkViz.git>`_ to your local machine:
+
+.. code-block:: bash
+
+    git clone https://github.com/PCMDI/ESMBenchmarkViz.git
+    cd ESMBenchmarkViz
+
+Step 2: (Optional) Create and Activate a Conda Environment
+----------------------------------------------------------
+
+It is recommended to use a `conda <https://docs.conda.io/en/latest/>`_ environment to manage dependencies. You can create and activate a new environment as follows:
+
+.. code-block:: bash
+
+    conda create -n esmbenchmarkviz -f conda_env/environment.yml
+    conda activate esmbenchmarkviz
+
+Step 3: Install the Package
+---------------------------
+
+Install the package using `pip` from the root directory of the cloned repository:
+
+.. code-block:: bash
+
+    pip install .
+
+Verifying the Installation
+--------------------------
+
+After installation, you can verify that the package is installed by running:
+
+.. code-block:: bash
+
+    python -c "import ESMBenchmarkViz; print(ESMBenchmarkViz.__version__)"
+
+If you see the version number printed without errors, the installation was successful.
+
+Troubleshooting
+---------------
+
[…]
