+++ _sources/generated/ESMBenchmarkViz.scatter_plot.rst.txt
@@ -1,0 +1,6 @@
+﻿ESMBenchmarkViz.scatter\_plot
+=============================
+
+.. currentmodule:: ESMBenchmarkViz
+
+.. autofunction:: scatter_plot
