+++ _sources/generated/ESMBenchmarkViz.portrait_plot.rst.txt
@@ -1,0 +1,6 @@
+﻿ESMBenchmarkViz.portrait\_plot
+==============================
+
+.. currentmodule:: ESMBenchmarkViz
+
+.. autofunction:: portrait_plot
