+++ .pre-commit-config.yaml
@@ -15,6 +15,7 @@ repos:
             args: [--markdown-linebreak-ext=md]
           - id: end-of-file-fixer
           - id: check-yaml
+            exclude: ^meta\.yaml$
 
     # =======================
     # Python code formatters
