### 2026-06-18 984797d
Update paper.md
