+++ README.md
@@ -31,12 +31,14 @@ Jiwoo Lee ([@lee1043](https://github.com/PCMDI/ESMBenchmarkViz/commits?author=le
 
 | <div style="width:300%">[Versions]</div> | Update summary   |
 | ------------- | ------------------------------------------- |
+| [v0.2.2]      | New placeholder images, Technical update
 | [v0.2.1]      | Technical update
 | [v0.2.0]      | Scatter plot and Portrait plot added
 | [v0.1.0]      | Taylor Diagram with diagnostic figure pop up capability
 | [v0.0.1]      | Initial release for interactive Taylor Diagram
 
 [Versions]: https://github.com/PCMDI/ESMBenchmarkViz/releases
+[v0.2.2]: https://github.com/PCMDI/ESMBenchmarkViz/releases/tag/v0.2.2
 [v0.2.1]: https://github.com/PCMDI/ESMBenchmarkViz/releases/tag/v0.2.1
 [v0.2.0]: https://github.com/PCMDI/ESMBenchmarkViz/releases/tag/v0.2.0
 [v0.1.0]: https://github.com/PCMDI/ESMBenchmarkViz/releases/tag/v0.1.0
