### 2026-07-02 3fdde74
Add citation for Hassler et al. (2026) article
Added a new article entry for Hassler et al. (2026) on climate model benchmarking.
