+++ docs/paper/paper.bib
@@ -118,3 +118,19 @@
    doi        = {10.5281/zenodo.10236521},
    url        = {https://doi.org/10.5281/zenodo.10236521}
 }
+
+@article{Hassler:2026,
+   author = {Hassler, Birgit and Hoffman, Forrest M. and Beadling, Rebecca and Blockley, Ed and Huang, Bo and Lee, Jiwoo and Lembo, Valerio and Lewis, Jared and Lu, Jianhua and Madaus, Luke and Malinina, Elizaveta and Medeiros, Brian and Pokam, Wilfried and Scoccimarro, Enrico and Swaminathan, Ranjini},
+   title = {Systematic Benchmarking of Climate Models: Methodologies, Applications, and New Directions},
+   journal = {Reviews of Geophysics},
+   volume = {64},
+   number = {1},
+   pages = {e2025RG000891},
+   keywords = {model evaluation, climate models, model benchmarking, CMIP},
+   doi = {https://doi.org/10.1029/2025RG000891},
+   url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2025RG000891},
+   eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2025RG000891},
+   note = {e2025RG000891 2025RG000891},
+   abstract = {Abstract As climate models become increasingly complex, there is a growing need to comprehensively and systematically assess model performance with respect to observations. Given the increasing number and diversity of climate model simulations in use, the community has moved beyond simple model intercomparison and toward developing methods capable of benchmarking a large number of simulations against a suite of climate metrics. Here, we present a detailed review of evaluation and benchmarking methods and approaches developed in the last decade, focusing primarily on scientific implications for Coupled Model Intercomparison Project (CMIP) simulations and CMIP6 results that contributed to the Intergovernmental Panel on Climate Change (IPCC) Sixth Assessment Report (AR6). Based on this review, we explain the resulting contemporary philosophy of model benchmarking, and provide clear distinctions and definitions of the terms model verification, process validation, evaluation, and benchmarking. While significant progress has been made in model development based on systematic evaluation and benchmarking efforts, some climate system biases still remain. The development of open-source community software packages has played a fundamental role in identifying areas of significant model improvement and bias reduction. We review the key features of several software packages that have been commonly used over the past decade to evaluate and benchmark global and regional climate models. Additionally, we discuss best practices for the selection of evaluation and benchmarking metrics and for interpreting the obtained results, the importance of selecting suitable sources of reference data and accurate uncertainty quantification.},
+   year = {2026}
+}
