### 2026-07-02 b106e96
Add citation to Earth System Models section
Added citation to support the importance of Earth System Models in climate prediction.
