### 2026-08-31 19eff89
fix invalid DOI prefix, following https://github.com/openjournals/joss-reviews/issues/10644#issuecomment-5480759354
