### 2026-08-31 PCMDI/ESMBenchmarkViz#46 @mengqi-z
JOSS Paper
Hi @lee1043,

This is for JOSS submission openjournals/joss-reviews#10644

Could you please address the following items in the paper? The line numbers are from the PDF version.

1. Line 30: `Bokeh` is a Python library, so this word should be formatted as code by using \`Bokeh\` in `paper.md`. This rule applies to all library names, model names, package names, etc. Please update the formatting for these terms throughout the paper, including Matplotlib, Plotly, Pandas, numpy, xarray, Python, etc.
2. Line 42: There should be parentheses around the citation, such as "(Lee et al., 2024)".
3. Line 43: The link `https://pcmdi.llnl.gov/research/metrics/` returns a "URL not found" error.
4. Figure 1 caption:
  * The citation for _Taylor, (2001)_ is not hyperlinked.
  * There is two parenthesis around _Gleckler et al., 2008_. Please remove one.
  * Please remove the word "model" after ESM since ESM already includes "model".
5. Line 98: The link has the same "URL not found" error.
6. Line 118: I suggest using "PCMDI" instead of the full name, since the abbreviation has already been defined.
7. Line 141: The RGMA link is not working.
8. Most citations don't need parentheses around the year. For example, the citation on line 13, _(Hassler et al. (2006))_ should be _(Hassler et al., 2006)_, with the entire citation being hyperlinked. A correct example from the paper is the citation on line 107: _(J. Lee et al., 2024)_. You only need to use the format _XYZ et al., (20XX)_ when the citation is part of the sentence. Thus, I believe all citations, except for the one on line 107, need to be corrected.
9. I also noticed that the author names in the references are formatted inconsistently. For example, _Jiwoo Lee et al. (2023)_ has only the first author's first name spelled out, while the other authors' first names are not. In contrast, _J. Lee et al. (2024)_ uses initials for all authors. Please make sure the author-name format is consistent throughout `paper.bib`.

Thanks!
