### 2026-08-31 PCMDI/ESMBenchmarkViz#46 @lee1043
Hi @mengqi-z, Thank you for opening the issue and your comments. While I am working on addressing these, it looks like some of reference related issues are resulted from rendering of the PDF file. Is there any way I could generate the PDF file locally for validation before pinging you for the final version?
