### 2026-08-31 200891a
fix following https://github.com/PCMDI/ESMBenchmarkViz/issues/46#issue-5303629553
