### 2026-08-31 c5fe30e
fix for https://github.com/PCMDI/ESMBenchmarkViz/issues/46#issue-5303629553
