### 2026-08-31 PCMDI/ESMBenchmarkViz#40 @lee1043
Closing this issue and follow https://github.com/PCMDI/ESMBenchmarkViz/issues/46 for JOSS paper review, while conda-forge release has been accomplished.
