### 2026-08-31 8f0d6b0
fix for https://github.com/PCMDI/ESMBenchmarkViz/issues/46#issue-5303629553
