### 2026-08-31 PCMDI/ESMBenchmarkViz#46 @mengqi-z
Hi @lee1043 - There are several ways to generate PDF without using `@editorialbot` in the review thread https://joss.readthedocs.io/en/latest/paper.html#locally.
