### 2026-08-31 596a45d
change paths for paper
