+++ .github/workflows/draft-pdf.yml
@@ -4,7 +4,6 @@ on:
     paths:
       - paper/**
       - .github/workflows/draft-pdf.yml
-  pull_request:
 
 jobs:
   paper:
@@ -26,4 +25,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper.pdf
+          path: docs/paper/paper.pdf
