+++ docs/paper/paper.md
@@ -46,7 +46,7 @@ Evaluating ESMs involves comparing model outputs against observational data and 
 
 # Statement of need
 
-ESM evaluations (particularly comprehensive evaluations) often generate large volumes of data (e.g., @Lee:2019, @Lee:2021, @Lee:2024, @Ahn:2022, @Planton:2021), which can be challenging to interpret and communicate effectively. Wrangling this data to produce an effective, polished visualization is a cumbersome process that often requires constructing data and label lists, calculating plot coordinates, and making manual adjustments for final elements like color bar position or label orientation.
+ESM evaluations (particularly comprehensive evaluations) often generate large volumes of data [e.g., @Lee:2019, @Lee:2021, @Lee:2024, @Ahn:2022, @Planton:2021], which can be challenging to interpret and communicate effectively. Wrangling this data to produce an effective, polished visualization is a cumbersome process that often requires constructing data and label lists, calculating plot coordinates, and making manual adjustments for final elements like color bar position or label orientation.
 
 To address these challenges, we have developed `ESMBenchmarkViz`, a modern Python library specifically designed for efficient interactive visualization of statistical performance metrics from ESM evaluation and intercomparison. This library leverages the power of `Bokeh` [@Bokeh:2018] to provide researchers and practitioners with user-friendly interactive tools for dynamic exploration of complex datasets. By enabling real-time manipulation of visualizations, `ESMBenchmarkViz` facilitates deeper insights into model performance and inter-model comparisons, making it easier to identify outliers and persistent biases.
 
@@ -54,17 +54,17 @@ In this document, we describe the core functionalities of `ESMBenchmarkViz` and 
 
 # Software Design
 
-The development of `ESMBenchmarkViz` originated from the interactive visualization dashboard of the [Program for Climate Model Diagnosis and Intercomparison (PCMDI) Metrics Package](https://pcmdi.github.io/pcmdi_metrics/) [@pcmdi-metrics] [@Lee:2024], showing diverse evaluation metrics for ESMs along with diagnostic information (https://pcmdi.llnl.gov/research/metrics/). We refer to these diagnostics as “dive-down information,” as they enable users to investigate metrics in greater detail.
+The development of `ESMBenchmarkViz` originated from the interactive visualization dashboard of the [`Program for Climate Model Diagnosis and Intercomparison (PCMDI) Metrics Package`](https://pcmdi.github.io/pcmdi_metrics/) [@Lee:2024], showing diverse evaluation metrics for ESMs along with diagnostic information (https://pcmdi.llnl.gov/research/metrics/). We refer to these diagnostics as “dive-down information,” as they enable users to investigate metrics in greater detail.
 
-`ESMBenchmarkViz` is structured as a modular `Python` library that provides high-level plotting functions tailored to common Earth system model (ESM) evaluation workflows. The core design goal is to minimize boilerplate for end users while retaining flexibility for advanced customization. To achieve this, the package separates three main concerns: (1) data preparation and validation, (2) plot configuration (layout, styling, and interactivity), and (3) rendering via the Bokeh backend. Each supported plot type (Taylor diagram, portrait plot, scatter plot) is implemented as a self-contained module exposing a concise, stable API, while sharing common utilities for color mapping, tooltips, legends, and export.
[…]
