+++ docs/paper/paper.bib
@@ -95,30 +95,6 @@
    url      = {https://bokeh.pydata.org/en/latest/},
 }
 
-
-@software{pcmdi-metrics,
-   author     = {Lee, Jiwoo
-               and Gleckler, Peter
-               and Ordonez, Ana
-               and Ahn, Min-Seop
-               and Ullrich, Paul
-               and Vo, Tom
-               and Boutte, Jason
-               and Doutriaux, Charles
-               and Durack, Paul
-               and Shaheen, Zeshawn
-               and Muryanto, Lina
-               and Painter, Jeffrey
-               and Krasting, John},
-   title      = {PCMDI/pcmdi\_metrics: PMP Version 3.1.2},
-   month      = nov,
-   year       = 2023,
-   publisher  = {Zenodo},
-   version    = {v3.1.2},
-   doi        = {10.5281/zenodo.10236521},
-   url        = {https://doi.org/10.5281/zenodo.10236521}
-}
-
 @article{Hassler:2026,
    author = {Hassler, Birgit and Hoffman, Forrest M. and Beadling, Rebecca and Blockley, Ed and Huang, Bo and Lee, Jiwoo and Lembo, Valerio and Lewis, Jared and Lu, Jianhua and Madaus, Luke and Malinina, Elizaveta and Medeiros, Brian and Pokam, Wilfried and Scoccimarro, Enrico and Swaminathan, Ranjini},
    title = {Systematic Benchmarking of Climate Models: Methodologies, Applications, and New Directions},
