### 2026-08-31 3758b54
lightweight build workflow (exclude when working on paper drafting)
