+++ .github/workflows/build_workflow.yml
@@ -6,6 +6,8 @@ on:
 
   pull_request:
     branches: [main]
+    paths-ignore:
+      - 'docs/paper/**'
 
   workflow_dispatch:
 
