+++ docs/paper/paper.bib
@@ -38,7 +38,7 @@
 }
 
 @article {Lee:2021,
-   author = "Jiwoo Lee and Kenneth R. Sperber and Peter J. Gleckler and Karl E. Taylor and Céline J. W. Bonfils",
+   author = "Jiwoo Lee, and Kenneth R. Sperber, and Peter J. Gleckler, and Karl E. Taylor, and Céline J. W. Bonfils",
    title = "Benchmarking Performance Changes in the Simulation of Extratropical Modes of Variability across CMIP Generations",
    journal = "Journal of Climate",
    year = "2021",
