+++ docs/paper/paper.md
@@ -46,7 +46,7 @@ Evaluating ESMs involves comparing model outputs against observational data and 
 
 # Statement of need
 
-ESM evaluations (particularly comprehensive evaluations) often generate large volumes of data, as examples shown in @Lee:2019, @Lee:2021, @Lee:2024, @Ahn:2022, and @Planton:2021, which can be challenging to interpret and communicate effectively. Wrangling this data to produce an effective, polished visualization is a cumbersome process that often requires constructing data and label lists, calculating plot coordinates, and making manual adjustments for final elements like color bar position or label orientation.
+ESM evaluations (particularly comprehensive evaluations) often generate large volumes of data, as examples shown in @Lee:2019, @Lee:2021, @Planton:2021, @Ahn:2022, and @Lee:2024, which can be challenging to interpret and communicate effectively. Wrangling this data to produce an effective, polished visualization is a cumbersome process that often requires constructing data and label lists, calculating plot coordinates, and making manual adjustments for final elements like color bar position or label orientation.
 
 To address these challenges, we have developed `ESMBenchmarkViz`, a modern Python library specifically designed for efficient interactive visualization of statistical performance metrics from ESM evaluation and intercomparison. This library leverages the power of `Bokeh` [@Bokeh:2018] to provide researchers and practitioners with user-friendly interactive tools for dynamic exploration of complex datasets. By enabling real-time manipulation of visualizations, `ESMBenchmarkViz` facilitates deeper insights into model performance and inter-model comparisons, making it easier to identify outliers and persistent biases.
 
@@ -60,7 +60,7 @@ The development of `ESMBenchmarkViz` originated from the interactive visualizati
 
 A key design trade-off was choosing `Bokeh` as the primary visualization engine rather than building on top of more general plotting wrappers. `Bokeh` offers native support for interactive HTML output, hover tools, and linked selections, which aligns with the “dive-down” paradigm that motivated this toolkit. Although other visualization libraries (e.g., `Matplotlib`, `Plotly`, `Holoviews`) provide overlapping capabilities, they either lack specialized implementations of ESM-oriented diagnostics (such as Taylor diagrams and portrait plots with standardized climate-centric defaults) or require substantial custom code to reproduce similar interactive features.
 
-In developing `ESMBenchmarkViz`, we considered contributing directly to existing packages, particularly the [`PCMDI Metrics Package`](https://pcmdi.github.io/pcmdi_metrics/) and general-purpose plotting libraries. We elected to build a standalone but interoperable toolkit for two reasons. First, we aimed for an architecture that is agnostic to any particular metrics engine, so that it can be reused across different ESM evaluation pipelines beyond PCMDI. Second, keeping visualization logic in a dedicated library allows us to iterate on interaction patterns, layout presets, and performance optimizations without constraining upstream packages or imposing additional dependencies on users whose workflows do not require interactive plots. At the same time, `ESMBenchmarkViz` is designed to integrate smoothly with existing climate analysis ecosystems through simple data structures (e.g., pure `Python` list, `Pandas` DataFrames, `numpy` or `xarray` outputs) and clear [APIs](https://pcmdi.github.io/ESMBenchmarkViz/api.html).
[…]
