+++ docs/paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Ahn:2022,
-   author = "Min-Seop Ahn and Peter J. Gleckler and Jiwoo Lee and Angeline G. Pendergrass and Christian Jakob",
+   author = {Min-Seop Ahn and Peter J. Gleckler and Jiwoo Lee and Angeline G. Pendergrass and Christian Jakob},
    title = "Benchmarking Simulated Precipitation Variability Amplitude across Time Scales",
    journal = "Journal of Climate",
    year = "2022",
