+++ docs/paper/paper.bib
@@ -8,7 +8,7 @@
    volume = "35",
    number = "20",
    doi = "10.1175/JCLI-D-21-0542.1",
-   pages=      "6773 - 6796",
+   pages = "6773 - 6796",
    url = "https://journals.ametsoc.org/view/journals/clim/35/20/JCLI-D-21-0542.1.xml"
 }
 
@@ -28,27 +28,27 @@
 }
 
 @article{Lee:2019,
-  author  = {Lee, J. and Sperber, K. and Gleckler, P. and Bonfils, C. and Taylor, K.},
-  title   = {Quantifying the Agreement Between Observed and Simulated Extratropical Modes of Interannual Variability},
-  journal = {Climate Dynamics},
-  year    = {2019},
-  volume  = {52},
-  pages   = {4057--4089},
-  doi     = {10.1007/s00382-018-4355-4}
+   author  = {Lee, J. and Sperber, K. and Gleckler, P. and Bonfils, C. and Taylor, K.},
+   title   = {Quantifying the Agreement Between Observed and Simulated Extratropical Modes of Interannual Variability},
+   journal = {Climate Dynamics},
+   year    = {2019},
+   volume  = {52},
+   pages   = {4057--4089},
+   doi     = {10.1007/s00382-018-4355-4}
 }
 
 @article {Lee:2021,
-   author = {Lee, J., and K. R. Sperber, and P. J. Gleckler, and K. E. Taylor, and C. Bonfils},
-   title = "Benchmarking Performance Changes in the Simulation of Extratropical Modes of Variability across CMIP Generations",
-   journal = "Journal of Climate",
-   year = "2021",
+   author = {Lee, J. and Sperber, K. R. and Gleckler, P. J. and Taylor, K. E. and Bonfils, C.},
+   title = {Benchmarking Performance Changes in the Simulation of Extratropical Modes of Variability across CMIP Generations},
+   journal = {Journal of Climate},
+   year = {2021},
    publisher = "American Meteorological Society",
    address = "Boston MA, USA",
-   volume = "34",
-   number = "17",
-   doi = "10.1175/JCLI-D-20-0832.1",
-   pages=      "6945 - 6969",
-   url = "https://journals.ametsoc.org/view/journals/clim/34/17/JCLI-D-20-0832.1.xml"
+   volume = {34},
+   number = {17},
+   doi = {10.1175/JCLI-D-20-0832.1},
+   pages = {6945--6969},
+   url = {https://journals.ametsoc.org/view/journals/clim/34/17/JCLI-D-20-0832.1.xml}
 }
 
 @Article{Lee:2024,
@@ -64,17 +64,17 @@
 }
 
 @article {Planton:2021,
-      author = "Yann Y. Planton and Eric Guilyardi and Andrew T. Wittenberg and Jiwoo Lee and Peter J. Gleckler and Tobias Bayr and Shayne McGregor and Michael J. McPhaden and Scott Power and Romain Roehrig and Jérôme Vialard and Aurore Voldoire",
-      title = "Evaluating Climate Models with the CLIVAR 2020 ENSO Metrics Package",
-      journal = "Bulletin of the American Meteorological Society",
-      year = "2021",
-      publisher = "American Meteorological Society",
-      address = "Boston MA, USA",
-      volume = "102",
-      number = "2",
-      doi = "10.1175/BAMS-D-19-0337.1",
-      pages=      "E193 - E217",
-      url = "https://journals.ametsoc.org/view/journals/bams/102/2/BAMS-D-19-0337.1.xml"
[…]
