### 2026-09-01 PCMDI/ESMBenchmarkViz#46 @lee1043
Hi @mengqi-z,

Thank you very much for your [comments](https://github.com/PCMDI/ESMBenchmarkViz/issues/46#issue-5303629553)! The latest revised manuscript is upload to the working branch (https://github.com/PCMDI/ESMBenchmarkViz/tree/joss-paper/docs/paper) and its PDF version is available here: [paper.pdf](https://github.com/user-attachments/files/31668556/paper.pdf)

The following is our point-to-point responses for your comments.

1.	Line 30: Bokeh is a Python library, so this word should be formatted as code by using `Bokeh` in paper.md. This rule applies to all library names, model names, package names, etc. Please update the formatting for these terms throughout the paper, including Matplotlib, Plotly, Pandas, numpy, xarray, Python, etc.
_**--> Corrected accordingly**_

2.	Line 42: There should be parentheses around the citation, such as "(Lee et al., 2024)".
**_--> Corrected accordingly_**

3.	Line 43: The link https://pcmdi.llnl.gov/research/metrics/ returns a "URL not found" error.
**_--> The link is working okay on my end. Could you please retry?_**

4.	Figure 1 caption:
•	The citation for Taylor, (2001) is not hyperlinked.
•	There is two parenthesis around Gleckler et al., 2008. Please remove one.
•	Please remove the word "model" after ESM since ESM already includes "model".
**_--> Corrected accordingly_**

5.	Line 98: The link has the same "URL not found" error.
**_--> The link is revised to be consistent to the above comment number 3, and I have double checked it is working okay on my end. Could you please confirm it works on your side as well?_**

6.	Line 118: I suggest using "PCMDI" instead of the full name, since the abbreviation has already been defined.
**_--> Corrected accordingly_**

7.	Line 141: The RGMA link is not working.
**_--> Corrected_**

8.	Most citations don't need parentheses around the year. For example, the citation on line 13, (Hassler et al. (2006)) should be (Hassler et al., 2006), with the entire citation being hyperlinked. A correct example from the paper is the citation on line 107: (J. Lee et al., 2024). You only need to use the format XYZ et al., (20XX) when the citation is part of the sentence. Thus, I believe all citations, except for the one on line 107, need to be corrected.
**_--> Corrected throughout the paper_**

9.	I also noticed that the author names in the references are formatted inconsistently. For example, Jiwoo Lee et al. (2023) has only the first author's first name spelled out, while the other authors' first names are not. In contrast, J. Lee et al. (2024) uses initials for all authors. Please make sure the author-name format is consistent throughout paper.bib.
**_--> Corrected throughout the paper_**

Thank you for your time!
