### 2026-09-01 PCMDI/ESMBenchmarkViz#46 @lee1043
> Hi [@lee1043](https://github.com/lee1043) - There are several ways to generate PDF without using `@editorialbot` in the review thread https://joss.readthedocs.io/en/latest/paper.html#locally.

This was very useful and helpful information, thank you for letting me know!
