### 2026-09-01 PCMDI/ESMBenchmarkViz#46 @mengqi-z
@lee1043 - Thank you for addressing my comments!

Regarding the “link not found” error I mentioned in No. 3, I found the cause. The URL itself is working, but when I directly click on the URL in the paper PDF to open the page, the URL somehow becomes `https://pcmdi.llnl.gov/research/metrics/)`, with an additional closing parenthesis added. This is what is causing the “URL not found” error.

I think this may be a syntax issue where the PDF recognizes the `)` as part of the URL. I believe if you update it to `[https://pcmdi.llnl.gov/research/metrics/](https://pcmdi.llnl.gov/research/metrics/)` in the paper, the correct link will be used.
