### 2026-09-01 PCMDI/ESMBenchmarkViz#46 @lee1043
Hi @mengqi-z, thank you for the investigation regarding the hyperlink! I've got them fixed and new PDF rendered version can be found here: [paper.pdf](https://github.com/user-attachments/files/31705427/paper.pdf)
