+++ docs/paper/paper.md
@@ -20,7 +20,7 @@ authors:
       orcid: 0009-0003-7608-0203
       affiliation: "1"
     - name: Peter Gleckler
-      orcid: 
+      orcid: 0000-0003-2816-6224
       affiliation: "1"
     - name: Paul Ullrich
       orcid: 0000-0003-4118-4590
