### 2026-09-01 #10644 @lee1043
Hi @mengqi-z, thank you for taking care of this and for your time and help!

- [x]  Double check authors and affiliations (including ORCIDs).
   > [Just added one missed ORCID](https://github.com/PCMDI/ESMBenchmarkViz/pull/47/changes/2ac94269bf4860f395f054d20c72f8e185f94940), sorry about that.

   > [v0.2.3](https://github.com/PCMDI/ESMBenchmarkViz/releases/tag/v0.2.3)

- [x] Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
   > https://zenodo.org/records/22217659

- [x] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
   > [Just added one missed ORCID](https://github.com/PCMDI/ESMBenchmarkViz/pull/47/changes/2ac94269bf4860f395f054d20c72f8e185f94940), sorry about that.

- [x] Please post the DOI of the archived version here.
   > https://doi.org/10.5281/zenodo.22217659
