### 2026-09-01 #10644 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
