### 2026-09-01 #10644 @mengqi-z
@lee1043 – I’ve recommended your submission for acceptance. Thanks so much for submitting to JOSS! From here, our Editor-in-Chief will step in to do some final checks before the paper can be officially accepted and published. Congrats and great job! 🎉

@shenyulu, @carbform – Huge thanks to all of you for reviewing this paper! I look forward to seeing your reviews for JOSS in the future!
