+++ ESMBenchmarkViz/core_taylor_diagram.py
@@ -388,7 +388,9 @@ def taylor_diagram(
             debug_print(debug, f"Static PNG exported to {static_filename}")
         except Exception as e:
             print(f"Failed to export PNG: {e}")
-            print("Tip: Install selenium and a browser driver (e.g., chromedriver) for PNG export")
+            print(
+                "Tip: Install selenium and a browser driver (e.g., chromedriver) for PNG export"
+            )
 
     # Show the plot if requested
     if show_plot:
