### 2026-09-03 fd4e6d9
Phases 1-2: Add consistency features to scatter_plot and portrait_plot
scatter_plot changes:
- Add export_png import and static PNG export
- Add show_image_panel parameter (default False)
- Add static, static_filename, bokeh_logo, debug parameters
- Add deepcopy protection for input lists
- Add 'save' tool to toolbar
- Improve image panel logic (optional now)
- Add error handling for PNG export

portrait_plot changes:
- Add deepcopy protection for img_url, xaxis_labels, yaxis_labels
- Improve PNG export error handling with try/except
- Add helpful tip for selenium/chromedriver installation

Br
[…]
