### 2026-09-03 1fca091
Fix portrait_plot: Reorder show() to match scatter_plot exactly
Changed order of operations to match working scatter_plot:
1. Export PNG (if static)
2. show() (if show_plot)
3. return

Was:
1. show() (if show_plot)
2. Export PNG (if static)
3. return

This matches the exact order in core_scatter_plot.py which works correctly.

Co-Authored-By: Claude <noreply@anthropic.com>
