+++ ESMBenchmarkViz/core_portrait_plot.py
@@ -501,7 +501,9 @@ def portrait_plot(
                 print(f"Static PNG exported to {static_filename}")
         except Exception as e:
             print(f"Failed to export PNG: {e}")
-            print("Tip: Install selenium and a browser driver (e.g., chromedriver) for PNG export")
+            print(
+                "Tip: Install selenium and a browser driver (e.g., chromedriver) for PNG export"
+            )
 
     # Show the plot if requested
     if show_plot:
