### 2026-09-03 91a82e0
Fix test_scatter_plot_with_images for new show_image_panel default
The scatter_plot function now has show_image_panel=False as default,
which means it returns a figure (not a layout) when images are provided.

Changes:
- test_scatter_plot_with_images: Updated to expect figure (default behavior)
- test_scatter_plot_with_images_and_panel: New test for layout (show_image_panel=True)

All 6 scatter_plot tests now pass.

Fixes: AttributeError: unexpected attribute 'children' to figure

Co-Authored-By: Claude <noreply@anthropic.com>
