### 2026-09-03 a895319
Add branch filter for draft PDF workflow
