+++ _sources/overview.rst.txt
@@ -1,0 +1,29 @@
+.. _overview:
+
+Overview
+========
+
+.. image:: _static/example_taylor_diagram.gif
+   :alt: Animated demonstration of the Taylor Diagram
+   :align: center
+
+Key Features
+------------
+* **Interactive Visualization**: ESMBenchmarkViz leverages Bokeh's capabilities to create dynamic and interactive plots that allow users to manipulate data views in real-time. This interactivity enhances the exploration of complex datasets, making it easier to identify trends and anomalies.
+* **Model Evaluation and Comparison**: The tool is particularly effective for evaluating outputs from multiple ESMs, allowing users to assess how well different models simulate various climate phenomena. This is crucial for understanding model strengths and weaknesses.
+* **Integration with Existing Frameworks**: ESMBenchmarkViz can be integrated with other Python libraries commonly used in data science, such as NumPy and Pandas. This compatibility allows for seamless data manipulation and analysis prior to visualization.
+* **User-Friendly Interface**: Designed with usability in mind, ESMBenchmarkViz provides an intuitive interface that simplifies the process of creating visualizations, making it accessible even for users with limited programming experience.
+
+Applications in Earth System Modeling
+-------------------------------------
+Earth System Models are essential tools for simulating climate dynamics and predicting future climate scenarios. The ability to benchmark these models against observations and each other is vital for improving their accuracy. ESMBenchmarkViz supports this process by:
+
+* **Facilitating Collaborative Research**: By enabling multiple researchers to visualize and compare model outputs collectively, the tool fosters collaboration within the scientific community.
+* **Enhancing Transparency in Model Evaluation**: Interactive visualizations help communicate model performance more effectively to stakeholders, including policymakers and the public, thereby enhancing transparency in climate science.
+* **Supporting Educational Initiatives**: The tool can also serve as an educational resource, helping students and new researchers understand complex climate modeling concepts through interactive learning experiences.
+
+In summary, ESMBenchmarkViz stands out as a powerful tool for visualizing and benchmarking Earth System Models. Its interactive features, combined with its integration capabilities and user-friendly design, make it a valuable asset for researchers aiming to enhance their understanding of climate dynamics through model evaluation.
+
+Reference
+---------
+Lee, J., K. Y. Chang, P. Gleckler, P. Ullrich, 2026: ESMBenchmarkViz: A Python Toolkit for Interactive Visualization of Earth System Model Evaluation and Benchmarking. Journal of Open Source Software (accepted)
