+++ _sources/examples/example_taylor_diagram_2.ipynb.txt
@@ -1,0 +1,209 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Interactive Taylor Diagram\n",
+    "\n",
+    "\n",
+    "![interactive_taylor_diagram](../_static/example_taylor_diagram.gif)\n",
+    "\n",
+    "\n",
+    "This notebook shows practical example use cases of interactive Taylor Diagram. Detailed API description can be found [here](../generated/ESMBenchmarkViz.taylor_diagram.html).\n",
+    "\n",
+    "## Import required functions"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from bokeh.plotting import output_file, save, output_notebook\n",
+    "from ESMBenchmarkViz import taylor_diagram"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/html": [
+       "    <style>\n",
+       "        .bk-notebook-logo {\n",
+       "            display: block;\n",
+       "            width: 20px;\n",
+       "            height: 20px;\n",
[…]
