+++ _sources/gallery.rst.txt
@@ -1,0 +1,12 @@
+.. _gallery:
+
+Gallery
+=======
+
+The following page contains a collection of practical examples visualized with ESMBenchmarkViz. 
+In addition, the user might come across tutorials related on how to use and customize the plotting utilities, insight on the `API <api.html>`_ capabilities.
+
+.. nbgallery::
+   examples/example_taylor_diagram
+   examples/example_portrait_plot
+   examples/example_scatter_plot
