### 2026-09-03 #10644 @lee1043
@kthyng thank you for reviewing this. As recommended, we've got the title of software and author list for the metadata in zenodo archive (https://zenodo.org/records/22217659) matching to the submitted paper.
