### 2026-09-03 #10644 @kthyng
@lee1043 Paper comments
- Page 1: “as examples shown in Lee et al. (2019)…” should be something like “as shown in Lee et al. (2019)…”
- Page 2: “pure Python list” should be “pure Python lists” to agree with the other plural items in the list (“Pandas DataFrames, numpy or xarray outputs”).
- Page 3, Figure 1 caption: should be something like, “A user’s mouse cursor hovering over a specific data point…” (extra "for");
  - also “a specific ESM and for its metrics” has an extraneous “for”: “a specific ESM and its metrics.”
  - also “Images can be also included to the tooltips or to the side viewer” should be “Images can also be included in the tooltips or in the side viewer.”
- Page 4: “as examples shown in Lee et al. (2019)…” should be “as shown in Lee et al. (2019)…”
- Page 4: “Although it is a very widely applied type of plot interdisciplinary” - "interdisciplinary" should presumably be replaced with “across disciplines” or similar
- Page 4: “we have included it to the package” should be “we have included it in the package.”
- Page 4: “tooltips and images accompanying together” - “Accompanying” appears to be extraneous
- Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, Lee et al. (2024) reference: “earth system models” should be “Earth system models”
