### 2026-09-03 #10644 @kthyng
> [@kthyng](https://github.com/kthyng) thank you for reviewing this. As recommended, we've got the title of software and author list for the metadata in zenodo archive (https://zenodo.org/records/22217659) matching to the submitted paper.

@lee1043 Thanks - though I'm expecting basically to see your long title ("ESMBenchmarkViz: A Python Toolkit for Interactive Visualization of Earth System Model Evaluation and Benchmarking") here instead of just "ESMBenchmarkViz" (again this is optional, but just to explain).
