### 2026-09-03 7a2ef8e
fix following https://github.com/openjournals/joss-reviews/issues/10644#issuecomment-5533208630
