+++ docs/paper/paper.bib
@@ -53,7 +53,7 @@
 
 @Article{Lee:2024,
    AUTHOR = {Lee, J. and Gleckler, P. J. and Ahn, M.-S. and Ordonez, A. and Ullrich, P. A. and Sperber, K. R. and Taylor, K. E. and Planton, Y. Y. and Guilyardi, E. and Durack, P. and Bonfils, C. and Zelinka, M. D. and Chao, L.-W. and Dong, B. and Doutriaux, C. and Zhang, C. and Vo, T. and Boutte, J. and Wehner, M. F. and Pendergrass, A. G. and Kim, D. and Xue, Z. and Wittenberg, A. T. and Krasting, J.},
-   TITLE = {Systematic and objective evaluation of Earth system models: PCMDI Metrics Package (PMP) version 3},
+   TITLE = {Systematic and Objective Evaluation of {Earth System Models}: PCMDI Metrics Package (PMP) version 3},
    JOURNAL = {Geoscientific Model Development},
    VOLUME = {17},
    YEAR = {2024},
