### 2026-09-03 #10644 @lee1043
@kthyng Thank you for the proofreading comments (https://github.com/openjournals/joss-reviews/issues/10644#issuecomment-5533208630), which I have fixed the manuscript following it.

Revised manuscript: [paper.pdf](https://github.com/user-attachments/files/31813510/paper.pdf)

Regarding the zenodo archive title, I appreciate the clarification, but I'd like to keep it as is for consistency to other versions as well as other zenodo archives under the institutional group.
